[CAP-88] Refactor payloads
