+++ src/main/java/org/capstone/job_fair/controllers/access_control/AuthController.java
@@ -6,8 +6,8 @@ import org.capstone.job_fair.config.jwt.JwtTokenProvider;
 import org.capstone.job_fair.config.jwt.details.UserDetailsImpl;
 import org.capstone.job_fair.constants.ApiEndPoint;
 import org.capstone.job_fair.constants.MessageConstant;
-import org.capstone.job_fair.controllers.payload.requests.LoginRequest;
-import org.capstone.job_fair.controllers.payload.requests.RefreshTokenRequest;
+import org.capstone.job_fair.controllers.payload.requests.accesss_control.LoginRequest;
+import org.capstone.job_fair.controllers.payload.requests.accesss_control.RefreshTokenRequest;
 import org.capstone.job_fair.controllers.payload.responses.LoginResponse;
 import org.capstone.job_fair.controllers.payload.responses.RefreshTokenResponse;
 import org.capstone.job_fair.models.entities.account.AccountEntity;
