+++ src/main/java/org/capstone/job_fair/controllers/access_control/ResetPasswordController.java
@@ -4,8 +4,8 @@ import lombok.extern.slf4j.Slf4j;
 import org.capstone.job_fair.constants.ApiEndPoint;
 import org.capstone.job_fair.constants.MessageConstant;
 import org.capstone.job_fair.constants.ResetPasswordTokenConstants;
-import org.capstone.job_fair.controllers.payload.requests.GenerateOTPRequest;
-import org.capstone.job_fair.controllers.payload.requests.ResetPasswordRequest;
+import org.capstone.job_fair.controllers.payload.requests.accesss_control.GenerateOTPRequest;
+import org.capstone.job_fair.controllers.payload.requests.account.ResetPasswordRequest;
 import org.capstone.job_fair.controllers.payload.responses.GenericResponse;
 import org.capstone.job_fair.models.entities.account.AccountEntity;
 import org.capstone.job_fair.models.entities.token.PasswordResetTokenEntity;
