+++ src/main/java/org/capstone/job_fair/controllers/attendant/AttendantController.java
@@ -2,8 +2,8 @@ package org.capstone.job_fair.controllers.attendant;
 
 import org.capstone.job_fair.constants.ApiEndPoint;
 import org.capstone.job_fair.constants.MessageConstant;
-import org.capstone.job_fair.controllers.payload.requests.RegisterAttendantRequest;
-import org.capstone.job_fair.controllers.payload.requests.UpdateAttendantRequest;
+import org.capstone.job_fair.controllers.payload.requests.attendant.RegisterAttendantRequest;
+import org.capstone.job_fair.controllers.payload.requests.attendant.UpdateAttendantRequest;
 import org.capstone.job_fair.controllers.payload.responses.GenericResponse;
 import org.capstone.job_fair.models.dtos.attendant.AttendantDTO;
 import org.capstone.job_fair.services.interfaces.account.AccountService;
