+++ src/main/java/org/capstone/job_fair/controllers/company/CompanyController.java
@@ -2,8 +2,8 @@ package org.capstone.job_fair.controllers.company;
 
 import org.capstone.job_fair.constants.ApiEndPoint;
 import org.capstone.job_fair.constants.MessageConstant;
-import org.capstone.job_fair.controllers.payload.requests.CreateCompanyRequest;
-import org.capstone.job_fair.controllers.payload.requests.UpdateCompanyRequest;
+import org.capstone.job_fair.controllers.payload.requests.company.CreateCompanyRequest;
+import org.capstone.job_fair.controllers.payload.requests.company.UpdateCompanyRequest;
 import org.capstone.job_fair.controllers.payload.responses.GenericResponse;
 import org.capstone.job_fair.models.dtos.company.CompanyDTO;
 import org.capstone.job_fair.services.interfaces.company.CompanyService;
