+++ src/main/java/org/capstone/job_fair/controllers/job_fair/LayoutController.java
@@ -4,8 +4,8 @@ import lombok.extern.slf4j.Slf4j;
 import org.capstone.job_fair.constants.AWSConstant;
 import org.capstone.job_fair.constants.ApiEndPoint;
 import org.capstone.job_fair.constants.MessageConstant;
-import org.capstone.job_fair.controllers.payload.requests.CreateLayoutMetaDataRequest;
-import org.capstone.job_fair.controllers.payload.requests.UpdateLayoutMetaDataRequest;
+import org.capstone.job_fair.controllers.payload.requests.job_fair.CreateLayoutMetaDataRequest;
+import org.capstone.job_fair.controllers.payload.requests.job_fair.UpdateLayoutMetaDataRequest;
 import org.capstone.job_fair.controllers.payload.responses.GenericResponse;
 import org.capstone.job_fair.models.dtos.job_fair.LayoutDTO;
 import org.capstone.job_fair.services.interfaces.job_fair.LayoutService;
