+++ src/main/java/org/capstone/job_fair/controllers/payload/requests/account/ResetPasswordRequest.java
@@ -1,4 +1,4 @@
-package org.capstone.job_fair.controllers.payload.requests;
+package org.capstone.job_fair.controllers.payload.requests.account;
 
 import lombok.*;
 import org.capstone.job_fair.validators.EmailConstraint;
