+++ src/main/java/org/capstone/job_fair/controllers/payload/requests/company/CreateCompanyRequest.java
@@ -1,15 +1,14 @@
-package org.capstone.job_fair.controllers.payload.requests;
+package org.capstone.job_fair.controllers.payload.requests.company;
 
 import lombok.*;
 import org.capstone.job_fair.constants.DataConstraint;
-import org.capstone.job_fair.models.statuses.CompanyStatus;
+import org.capstone.job_fair.constants.MessageConstant;
 import org.capstone.job_fair.validators.EmailConstraint;
 import org.capstone.job_fair.validators.PhoneConstraint;
 import org.capstone.job_fair.validators.XSSConstraint;
-import org.hibernate.validator.constraints.Length;
 
 import javax.validation.Valid;
-import javax.validation.constraints.Min;
+import javax.validation.constraints.NotBlank;
 import javax.validation.constraints.NotNull;
 import javax.validation.constraints.Size;
 import java.util.List;
@@ -19,36 +18,34 @@ import java.util.List;
 @AllArgsConstructor
 @Getter
 @Setter
-public class UpdateCompanyRequest {
-    @NotNull
-    private String id;
-
-    @Length(min = DataConstraint.Company.TAX_ID_LENGTH, max = DataConstraint.Company.TAX_ID_LENGTH)
+@ToString
+public class CreateCompanyRequest {
+    @NotBlank(message = MessageConstant.InvalidFormatValidationMessage.NOT_BLANK_FORMAT)
     private String taxId;
 
-    @Size(min = DataConstraint.Company.NAME_MIN_LENGTH, max = DataConstraint.Company.NAME_MAX_LENGTH)
-    @XSSConstraint
+    @NotNull
+    @Size(max = DataConstraint.Company.NAME_MAX_LENGTH)
     private String name;
 
-    @Size(min = DataConstraint.Company.ADDRESS_MIN_LENGTH, max = DataConstraint.Company.ADDRESS_MAX_LENGTH)
+    @NotNull
+    @Size(max = DataConstraint.Company.ADDRESS_MAX_LENGTH)
     @XSSConstraint
     private String address;
 
+    @NotNull
     @PhoneConstraint
     private String phone;
 
+    @NotNull
     @EmailConstraint
+    @XSSConstraint
     private String email;
-
-    @Min(value = DataConstraint.Company.COMPANY_MIN_NUM)
-    private Integer employeeMaxNum;
 
     @XSSConstraint
     private String url;
 
-    private Integer sizeId;
 
-    private CompanyStatus status;
+    private Integer sizeId;
 
     @Valid
     private List<String> mediaUrls;
@@ -67,4 +64,5 @@ public class UpdateCompanyRequest {
         private Integer id;
         private String description;
     }
+
 }
