+++ src/main/java/org/capstone/job_fair/services/interfaces/company/CompanyEmployeeService.java
@@ -3,6 +3,7 @@ package org.capstone.job_fair.services.interfaces.company;
 import org.capstone.job_fair.models.dtos.company.CompanyEmployeeDTO;
 
 import java.util.List;
+import java.util.Optional;
 
 public interface CompanyEmployeeService {
     CompanyEmployeeDTO createNewCompanyManagerAccount(CompanyEmployeeDTO dto);
@@ -18,5 +19,7 @@ public interface CompanyEmployeeService {
     void updateEmployeeStatus(String email);
 
     void promoteEmployee(String employeeId, String managerId);
+
+    Optional<CompanyEmployeeDTO> getCompanyEmployeeByAccountId(String accountID);
 
 }
