+++ src/main/java/org/capstone/job_fair/services/interfaces/company/CompanyRegistrationService.java
@@ -15,7 +15,5 @@ public interface CompanyRegistrationService {
 
     void staffEvaluateCompanyRegistration(String companyRegistrationId, CompanyRegistrationStatus status, String message);
 
-    String getCompanyIdInSecurityContext();
-
-    List<CompanyRegistrationDTO> getAllOwnCompanyRegistrationOfAJobFair(String jobFairId, String companyId);
+    List<CompanyRegistrationDTO> getCompanyRegistrationByJobFairIDAndCompanyID(String jobFairId, String companyId);
 }
