+++ src/main/java/org/capstone/job_fair/services/mappers/attendant/SkillMapper.java
@@ -1,6 +1,6 @@
 package org.capstone.job_fair.services.mappers.attendant;
 
-import org.capstone.job_fair.controllers.payload.requests.UpdateAttendantRequest;
+import org.capstone.job_fair.controllers.payload.requests.attendant.UpdateAttendantRequest;
 import org.capstone.job_fair.models.dtos.attendant.cv.SkillDTO;
 import org.capstone.job_fair.models.entities.attendant.cv.SkillEntity;
 import org.mapstruct.Mapper;
