+++ src/main/java/org/capstone/job_fair/services/mappers/attendant/WorkHistoryMapper.java
@@ -1,7 +1,7 @@
 package org.capstone.job_fair.services.mappers.attendant;
 
 
-import org.capstone.job_fair.controllers.payload.requests.UpdateAttendantRequest;
+import org.capstone.job_fair.controllers.payload.requests.attendant.UpdateAttendantRequest;
 import org.capstone.job_fair.models.dtos.attendant.cv.WorkHistoryDTO;
 import org.capstone.job_fair.models.entities.attendant.cv.WorkHistoryEntity;
 import org.mapstruct.Mapper;
