+++ src/main/java/org/capstone/job_fair/services/mappers/company/CompanyEmployeeMapper.java
@@ -1,6 +1,6 @@
 package org.capstone.job_fair.services.mappers.company;
 
-import org.capstone.job_fair.controllers.payload.requests.CompanyEmployeeRegisterRequest;
+import org.capstone.job_fair.controllers.payload.requests.company.CompanyEmployeeRegisterRequest;
 import org.capstone.job_fair.models.dtos.company.CompanyEmployeeDTO;
 import org.capstone.job_fair.models.entities.company.CompanyEmployeeEntity;
 import org.capstone.job_fair.services.mappers.account.AccountMapper;
